2025.2.24(day3)

+++ online-study-content/online-study-content-model/src/main/java/com/online_study/model/po/CourseMarket.java
@@ -16,9 +16,9 @@ public class CourseMarket {
 
     private String charge;
 
-    private Double price;
+    private Float price;
 
-    private Double originalPrice;
+    private Float originalPrice;
 
     private String qq;
